requirements.txt
flask==2.2.3
sqlalchemy==2.0.5.post1
sqlalchemy-utils==0.40.0
flask-mysqldb==1.0.1
pymysql==1.0.2
python-decouple==3.8
gunicorn==20.1.0
